[ updated ] refactor class
