+++ src/main/scala/com/bigchange/util/HBaseUtil.scala
@@ -1,13 +1,14 @@
 package com.bigchange.util
 
+import com.bigchange.hbase.HBase
 import com.bigchange.log.CLogger
+import org.apache.hadoop.hbase.TableName
 import org.apache.hadoop.hbase.client._
 import org.apache.hadoop.hbase.io.ImmutableBytesWritable
 import org.apache.hadoop.hbase.mapreduce.TableInputFormat
 import org.apache.hadoop.hbase.protobuf.ProtobufUtil
 import org.apache.hadoop.hbase.protobuf.generated.ClientProtos
 import org.apache.hadoop.hbase.util.{Base64, Bytes}
-import org.apache.hadoop.hbase.{HBaseConfiguration, HColumnDescriptor, HTableDescriptor, TableName}
 import org.apache.spark.SparkContext
 
 import scala.collection.mutable
@@ -15,7 +16,7 @@ import scala.collection.mutable
 /**
   * Created by C.J.YOU on 2016/3/21.
   */
-object HBaseUtil extends CLogger {
+object HBaseUtil extends HBase with CLogger {
 
   private val hBaseConfiguration = getConfiguration
 
@@ -25,81 +26,9 @@ object HBaseUtil extends CLogger {
 
   // 自定义外部可访问接口：存储对应格式数据到hbase即可（部分对象 特征 可根据实际需求 再进行一层抽象）
 
-  private def getConfiguration = {
-
-    // 本地测试 需配置的选项， 在集群上通过对应配置文件路径自动获得
-    val hbaseConf = HBaseConfiguration.create
-
-    hbaseConf.set("fs.defaultFS", "hdfs://ns1"); // nameservices的路径
-    hbaseConf.set("dfs.nameservices", "ns1");  //
-    hbaseConf.set("dfs.ha.namenodes.ns1", "nn1,nn2"); //namenode的路径
-    hbaseConf.set("dfs.namenode.rpc-address.ns1.nn1", "server3:9000"); // namenode 通信地址
-    hbaseConf.set("dfs.namenode.rpc-address.ns1.nn2", "server4:9000"); // namenode 通信地址
-    // 设置namenode自动切换的实现类
-    hbaseConf.set("dfs.client.failover.proxy.provider.ns1", "org.apache.hadoop.hdfs.server.namenode.ha.ConfiguredFailoverProxyProvider")
-    hbaseConf.set("hbase.rootdir", "hdfs://ns1/hbase")
-    hbaseConf.set("hbase.zookeeper.quorum", "server0,server1,server2")
-    hbaseConf.set("hbase.zookeeper.property.clientPort", "2181")
-
-    hbaseConf
-
-  }
-  /**
-    * 获取hbase的连接器
-    * @return connection
-    */
-  private def getConnection: Connection = {
-
-    warn("create connection")
-
-    val connection = ConnectionFactory.createConnection(hBaseConfiguration)
-
-    sys.addShutdownHook {
-      connection.close ()
-    }
-    connection
-  }
-
-  /**
-    * 创建hbase表
-    * @param tableName 表名
-    * @param columnFamilys 列族的声明
-    * @param connection 连接器
-    */
-  private def  createHbaseTable(tableName: TableName,columnFamilys:List[String],connection: Connection): Unit ={
-   // connection.getAdmin.createTable(new HTableDescriptor(tableName).addFamily(new HColumnDescriptor(columnFamily).setMaxVersions(3)))
-    val admin = connection.getAdmin
-    val htd = new HTableDescriptor(tableName)
-    for (family <- columnFamilys) {
-      val hcd = new HColumnDescriptor(family)
-      htd.addFamily(hcd.setMaxVersions(3))
-    }
-    admin.createTable(htd)
-    admin.close()
-  }
-
-  /**
-    * 判断 row key 是否存在
-    * @param row rowKey
-    * @param table tableName
-    * @return Boolean
-    */
-  private def existRowKey(row:String,table:Table): Boolean ={
-
-    val get = new Get(row.getBytes())
-    val result = table.get(get)
-
-    if (result.isEmpty) {
-      warn("hbase table don't have this data,execute insert")
-      return false
-    }
-
-    true
-
-  }
-
   /**
     * 数据格式转换
+    *
     * @param qualifiers  对应的列名集合
     * @param string 写入的数据内容
     * @return 自定义数据格式
@@ -118,6 +47,7 @@ object HBaseUtil extends CLogger {
 
   /**
     * 数据写入table中
+    *
     * @param tableName 表名
     * @param family  列族
     * @param qualifiers RowTelecomData 格式数据
@@ -149,7 +79,7 @@ object HBaseUtil extends CLogger {
 
   /**
     * 通过rowKey 获取数据
-    * @param rowKey  键
+    * @param rowKey 键
     * @param tableName 表名
     * @param family 列族
     * @param qualifiers 列名集合
@@ -170,6 +100,7 @@ object HBaseUtil extends CLogger {
 
   /**
     * 通过rdd 形式批量获取hbase数据
+    *
     * @param sc sparkContext
     * @param property 获取数据的属性：tableName, family, qualifiers , ...
     * @param scan  scan 定义
@@ -185,7 +116,7 @@ object HBaseUtil extends CLogger {
       hBaseConfiguration.set(key, value)
     }
 
-    val hbaseRDD = sc.newAPIHadoopRDD(hBaseConfiguration, classOf[TableInputFormat], classOf[ImmutableBytesWritable], classOf[Result]).map(_._2)
+    val hbaseRDD = sc.newAPIHadoopRDD(hBaseConfiguration, classOf[TableInputFormat], classOf[ImmutableBytesWritable], classOf[Result]).map(_._2 )
 
     hbaseRDD
